requirements.txt
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.2
django-taggit==1.2.0
gunicorn==20.0.4
psycopg2==2.8.4
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
static3==0.7.0
